+++ CoreData/Classes/SMPredicate.h
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-#import <Foundation/Foundation.h>
 #import <CoreLocation/CoreLocation.h>
+#import "SMGeoPoint.h"
 
 #define GEOQUERY_FIELD @"field"
 #define GEOQUERY_MILES @"miles"
@@ -34,9 +34,9 @@ typedef enum {
 } SMPredicateOperatorType;
 
 /**
- `SMPredicate` is a subclass of NSPredicate, with additional methods to build predicates that are specific to the StackMob API.  
+ `SMPredicate` is a subclass of `NSPredicate`, with additional methods to build predicates that are specific to the StackMob API.  
  
- @note Fetching from the cache using SMPredicate is not supported, and will return an empty array of results.
+ **Important:** Fetching from the cache using `SMPredicate` is not supported, and will return an empty array of results. Similarly, when a fetch is performed from the network (StackMob), any results are not cached.
  
  
  ## References ##
@@ -51,10 +51,6 @@ typedef enum {
 
 
 #pragma mark - Where clauses
-///-------------------------------
-/// @name Geo Location Clauses
-///-------------------------------
-
 /**
  Add the predicate criteria: `field`'s location is within `miles` of `point`.
  
@@ -66,8 +62,20 @@ typedef enum {
  
  @return An SMPredicate instance ready to be added to a fetch request.
  */
++ (SMPredicate *)predicateWhere:(NSString *)field isWithin:(CLLocationDistance)miles milesOf:(CLLocationCoordinate2D)point;
 
-+(SMPredicate *)predicateWhere:(NSString *)field isWithin:(CLLocationDistance)miles milesOf:(CLLocationCoordinate2D)point;
+/**
+ Add the predicate criteria: `field`'s location is within `miles` of `geoPoint`.
+ 
+ StackMob will generate a field `distance` and insert it into the response. This field is the distance between the query `field`'s location and `geoPoint`.
+ 
+ @param field The geo field in the StackMob schema that is to be compared.
+ @param miles Distance in miles.
+ @param geoPoint The SMGeoPoint around which to search.
+ 
+ @return An SMPredicate instance ready to be added to a fetch request.
+ */
++ (SMPredicate *)predicateWhere:(NSString *)field isWithin:(CLLocationDistance)miles milesOfGeoPoint:(SMGeoPoint *)geoPoint;
 
 /**
  Add the predicate criteria: `field`'s location is within `kilometers` of `point`.
@@ -80,8 +88,20 @@ typedef enum {
  
  @return An SMPredicate instance ready to be added to a fetch request.
  */
++ (SMPredicate *)predicateWhere:(NSString *)field isWithin:(CLLocationDistance)kilometers kilometersOf:(CLLocationCoordinate2D)point;
 
-+(SMPredicate *)predicateWhere:(NSString *)field isWithin:(CLLocationDistance)kilometers kilometersOf:(CLLocationCoordinate2D)point;
+/**
+ Add the predicate criteria: `field`'s location is within `kilometers` of `geoPoint`.
+ 
+ StackMob will generate a field `distance` and insert it into the response. This field is the distance between the query `field`'s location and `geoPoint`.
+ 
+ @param field The geo field in the StackMob schema that is to be compared.
+ @param kilometers Distance in kilometers.
+ @param geoPoint The SMGeoPoint around which to search.
+ 
+ @return An SMPredicate instance ready to be added to a fetch request.
+ */
++ (SMPredicate *)predicateWhere:(NSString *)field isWithin:(CLLocationDistance)kilometers kilometersOfGeoPoint:(SMGeoPoint *)geoPoint;
 
 /**
  Add the predicate criteria: `field`'s location falls within the bounding box with corners `sw` and `ne`.
@@ -92,22 +112,45 @@ typedef enum {
  
  @return An SMPredicate instance ready to be added to a fetch request.
  */
++ (SMPredicate *)predicateWhere:(NSString *)field isWithinBoundsWithSWCorner:(CLLocationCoordinate2D)sw andNECorner:(CLLocationCoordinate2D)ne;
 
-+(SMPredicate *)predicateWhere:(NSString *)field isWithinBoundsWithSWCorner:(CLLocationCoordinate2D)sw andNECorner:(CLLocationCoordinate2D)ne;
+/**
+ Add the predicate criteria: `field`'s location falls within the bounding box with corners `sw` and `ne`.
+ 
+ @param field The geo field in the StackMob schema that is to be compared.
+ @param sw SMGeoPoint of the bounding box's southwest corner.
+ @param ne SMGeoPoint of the bounding box's northeast corner.
+ 
+ @return An SMPredicate instance ready to be added to a fetch request.
+ */
++ (SMPredicate *)predicateWhere:(NSString *)field isWithinBoundsWithSWGeoPoint:(SMGeoPoint *)sw andNEGeoPoint:(SMGeoPoint *)ne;
 
 /**
  Add the predicate criteria: `field`'s location near point.
  
- StackMob will insert a field `distance` and insert it into the response. This field is the distance between the query `field`'s location and `location`.
+ StackMob will insert a field `distance` and insert it into the response. This field is the distance between the query `field`'s location and `point`.
  
  @note You probably want to apply a limit when including this predicate or you may end up with more results than intended.
  
  @param field The geo field in the StackMob schema that is to be compared.
- @param location The reference location.
+ @param point The reference location.
  
  @return An SMPredicate instance ready to be added to a fetch request.
  */
++ (SMPredicate *)predicateWhere:(NSString *)field near:(CLLocationCoordinate2D)point;
 
-+(SMPredicate *)predicateWhere:(NSString *)field near:(CLLocationCoordinate2D)point;
+/**
+ Add the predicate criteria: `field`'s location near point.
+ 
+ StackMob will insert a field `distance` and insert it into the response. This field is the distance between the query `field`'s location and `geoPoint`.
+ 
+ @note You probably want to apply a limit when including this predicate or you may end up with more results than intended.
+ 
+ @param field The geo field in the StackMob schema that is to be compared.
+ @param geoPoint The reference SMGeoPoint.
+ 
+ @return An SMPredicate instance ready to be added to a fetch request.
+ */
++ (SMPredicate *)predicateWhere:(NSString *)field nearGeoPoint:(SMGeoPoint *)geoPoint;
 
 @end