+++ integration tests/CRUDTests.m
@@ -19,6 +19,7 @@
 
 SPEC_BEGIN(CRUDTests)
 
+
 describe(@"CRUD", ^{
     __block SMDataStore *dataStore = nil;
     beforeEach(^{
@@ -244,6 +245,53 @@ describe(@"CRUD", ^{
             [[returnedSchema should] equal:@"Book"];
         });
 
+    });
+});
+
+describe(@"CRUD with GeoPoints", ^{
+    __block SMDataStore *dataStore = nil;
+    __block SMGeoPoint *point = nil;
+    __block NSDictionary *object = nil;
+    
+    beforeEach(^{;
+        dataStore = [SMIntegrationTestHelpers dataStore];
+        [NSURLRequest setAllowsAnyHTTPSCertificate:YES forHost:@"api.stackmob.com"];
+        
+        // Fisherman's Wharf
+        NSNumber *lat = [NSNumber numberWithDouble:37.810317];
+        NSNumber *lon = [NSNumber numberWithDouble:-122.418167];
+        point = [SMGeoPoint geoPointWithLatitude:lat longitude:lon];
+        
+        NSDictionary *args = [NSDictionary dictionaryWithObjectsAndKeys:@"StackMob", @"name", point, @"geopoint", nil];
+        
+        syncWithSemaphore(^(dispatch_semaphore_t semaphore) {
+            [dataStore createObject:args inSchema:@"random" onSuccess:^(NSDictionary *theObject, NSString *schema) {
+                object = theObject;
+                syncReturn(semaphore);
+            } onFailure:^(NSError *theError, NSDictionary *theObject, NSString *schema) {
+                syncReturn(semaphore);
+            }];
+        });
+    });
+    afterEach(^{
+        syncWithSemaphore(^(dispatch_semaphore_t semaphore) {
+            [dataStore deleteObjectId:[object objectForKey:@"random_id"] inSchema:@"random" onSuccess:^(NSString *theObjectId, NSString *schema) {
+                syncReturn(semaphore);
+            } onFailure:^(NSError *theError, NSString *theObjectId, NSString *schema) {
+                syncReturn(semaphore);
+            }];
+        });
+        object = nil;
+    });
+    it(@"Saves SMGeoPoint without error", ^{
+        [object shouldNotBeNil];
+    });
+    it(@"Reads SMGeoPoints correctly", ^{
+        
+        SMGeoPoint *geopoint = [object objectForKey:@"geopoint"];
+        
+        [[geopoint.latitude should] equal:point.latitude];
+        [[geopoint.longitude should] equal:point.longitude];
     });
 });
 
@@ -532,5 +580,6 @@ describe(@"setExpandDepth", ^{
     
     
 });
+
 
 SPEC_END